Claim the crates the generated Rust reaches, and state one fact once
`alloc` is declared with `extern crate` at the crate root under `no_std`, so
a DSDL namespace component of that name was a second root declaration and
E0260. The std profile compiles either way, which is why nothing had found
it. `core` and `std` join it: the bodies are written in terms of those
paths, and a module of either name would answer for them rather than the
crate. All three are claimed beside the runtime's own modules.

Which languages report a type name the manifest's consumer can reach the
type by was stated twice -- once in the manifest, once in the naming golden
-- and the two had already drifted apart in this branch. It is a property of
how a language composes a definition's name, so it is a field of
`DefinitionNamePolicy` and both read it. The values are the ones in force.
The reason once given for excluding C++ -- that its emitter builds a
namespace-qualified symbol of its own -- is not what `cppTypeName` does, and
whether C++ should report is left to the phase that takes C++.

The redundant cast clang-tidy reported is gone; `depth` is already a
`std::size_t`.

+++ lib/CodeGen/emitter/Rust.cpp
@@ -261,8 +261,7 @@ public:
         for (std::size_t depth = 1; taken(local) && (depth <= ref.namespaceComponents.size()); ++depth)
         {
             std::string raw;
-            for (const auto& component :
-                 llvm::ArrayRef<std::string>(ref.namespaceComponents).take_back(static_cast<std::size_t>(depth)))
+            for (const auto& component : llvm::ArrayRef<std::string>(ref.namespaceComponents).take_back(depth))
             {
                 raw += component + "_";
             }
